requirements.txt
streamlit==1.27.0
requests==2.31.0
pandas==2.0.3
python-dotenv==1.0.0
streamlit-option-menu==0.3.6
arabicreshaper==3.0.0
python-bidi==0.4.2
arabic-reshaper==3.0.0
python-bidi==0.4.2
reportlab==3.6.13
xhtml2pdf==0.2.11
